backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
requests
sqlalchemy
chromadb
sentence-transformers
aiofiles
numpy
